Broke distance functions out into separate file (except for conics; will do that later) and then added distance functions for corners.   Also tried a different disambiguating rule for determining the sign (which unfortunately doesn't quite work, and I don't know why, yet).   Added new test shaders for each of these.
Unfortunately, "preshrinking" the edges still seems to be the
fastest approach.   Unfortunately it will not be robust when
we quantize edges (IF we preshrink.  Could shrink the parameter
domain in the shader.. but then it won't be as fast).

BTw, this is Michael... working under Zheng's account, trying
to check things in and get things back under my account.   Really
wish we had two separate home machines to work on....


git-svn-id: svn://svn.code.sf.net/p/libsh/code/shrike/branches/vectortexture@2254 afdca40c-03d6-0310-8ede-e9f093b21075